feat: counter effect, embed timing fix, font resolution, new test files
- Implement counter effect (com.alightcreative.effects.counter)
  - Extract offset/scale params, animate text number transformation
  - Match AM's JS regex and toFixed behavior
- Fix embed scene timing: change time_offset from i32 to f32
  - Prevents truncation of fractional half-frame offset (16.667ms)
  - Add half-frame offset in spawn_embed.rs (same as collect_embed.rs)
- Fix video comparison time: floor to integer ms matching AM's frameStartTimeFromFrameNumber
- Fix font resolution: add Debian/Ubuntu unhinted font path, fix fc-match style matching
- Install Roboto font path for proper text rendering
- Add new test files: jetter (6), rgb-split (12), exposure-gamma (5),
  copy-bg (2), enlarge-bg (8), blend (5), chroma-key (1)
- Fix clippy issues: collapse if-let, extract transform_number helper, strip_prefix

+++ src/animation/spawn.rs
@@ -40,7 +40,7 @@ pub(crate) fn process_pending_layers(
     white_pixel: Option<&Handle<Image>>,
     global_time: f32,
     parent_entity: Entity,
-    time_offset: i32,
+    time_offset: f32,
     filter: &crate::scene::LayerFilter,
 ) {
     // We need to collect actions to avoid borrowing issues
@@ -52,7 +52,7 @@ pub(crate) fn process_pending_layers(
         layer_id: u64,
         layers: &[PendingLayer],
         global_time: f32,
-        _time_offset: i32,
+        _time_offset: f32,
     ) -> bool {
         is_ancestor_active_impl(layer_id, layers, global_time, _time_offset, &mut Vec::new())
     }
@@ -61,7 +61,7 @@ pub(crate) fn process_pending_layers(
         layer_id: u64,
         layers: &[PendingLayer],
         global_time: f32,
-        _time_offset: i32,
+        _time_offset: f32,
         visited: &mut Vec<u64>,
     ) -> bool {
         // Cycle detection: if we've already visited this layer, assume active to break cycle
